Set fragile and specific package-lock.json to avoid build errors
The upstream package does not build with node version >=12.8
due to some errors in babel. @babel/core and preset-env must be
pinned to ^7.9.0, which as of the date of this commit pulls in
version 7.10.2. This version is confirmed to work.

Just erasing the lock file and node_modules/ and running `npm install`
does **not** work, because that upgrades all packages and dependencies,
including some package that prevents several TypeScript files from
transpiling wihout errors. It's not the typescript package, I tried
holding that back to upstream's version and the errors were still
raised. I have no idea which package it is, there are nearly 2800
of them. This is why Javascript development can be so hellish.

The transpilation errors appear legitimate, so the real root cause
is non-conformant code in dash-table that wasn't caught by linting.

The patch is to run `npm ci` against upstream's `package-lock.json`,
then to update only the two babel packages that need it:

    npm install @babel/core@7.10.2
    npm install @babel/preset-env@7.10.2

The package-lock.json file in this commit works, meaning that
running `npm ci` and `npm run build` results in no errors. Of
course, `npm run build` has to be run from `dash`, which is
why the dash project's circleci is the single source of truth
regarding build instructions (can be done locally in a circleci
docker image, but according to dash's `.circleci/config.yml`).

Fragility squared. Lots of tech debt.